requirements.txt
flet
openai
requests
python-dotenv
spacy
nltk
numpy
pandas
scikit-learn